Split new module Symbolic.hs from Data.hs, containing low-level operations
Symbolic.hs defines types Op, SW, SBVExpr, Symbolic a, and related types.
Data.hs defines type SBV a, and various type classes.

Symbolic.hs exports new variants of a few old functions, with types that
use SVal instead of SBool. We also introduce a non-parameterized version
of 'SArray a b' called 'SArr', along with a set of basic operations.
